Work in progress 	modified:   src/HTTP.jl 	modified:   src/WebSockets.jl 	modified:   test/REQUIRE 	copied:     test/client_server_test.jl -> test/client_server_functions.jl 	modified:   test/client_server_test.jl 	deleted:    test/tempcoroutines.jl
