+++ test/REQUIRE
@@ -1,2 +1,3 @@
 WebSockets
 BufferedStreams
+Suppressor
